Refactored code to separate models for all screen
