+++ SwiftCubits/Manager/AuthManager.swift
